+++ comic_dl/comic_dl.py
@@ -22,6 +22,8 @@ class ComicDL(object):
         parser.add_argument('-rn', '--range', nargs=1, help='Specifies the range of chapters to download.', default='All')
         parser.add_argument('--convert', nargs=1, help='Tells the script to convert the downloaded Images to PDF or anything else.')
         parser.add_argument('--keep', nargs=1, help='Tells the script whether to keep the files after conversion or not.', default='True')
+        parser.add_argument('--quality', nargs=1,
+                            help='Tells the script which Quality of image to download (High/Low).', default='True')
 
         parser.add_argument('-i', '--input', nargs=1, help='Inputs the URL to anime.')
 
@@ -64,10 +66,12 @@ class ComicDL(object):
                 args.convert = ["None"]
             if not args.keep:
                 args.keep = ["True"]
+            if not args.keep:
+                args.keep = ["Best"]
 
             start_time = time.time()
             honcho.Honcho().checker(comic_url=str(args.input[0]).strip(), current_directory=os.getcwd(),
-                                    sorting_order=args.sorting[0], logger=logger, download_directory=args.download_directory[0], chapter_range=args.range, conversion=args.convert[0], delete_files=args.keep[0])
+                                    sorting_order=args.sorting[0], logger=logger, download_directory=args.download_directory[0], chapter_range=args.range, conversion=args.convert[0], delete_files=args.keep[0], image_quality=args.quality[0])
             end_time = time.time()
             total_time = end_time - start_time
             print("Total Time Taken To Complete : %s" % total_time)